Added autoinit feature to malloc so it can be used before main(). Changed build.sh script to print only errors and warnings. Fixed some warnings in math functions.

+++ src/math/pow.c
@@ -126,7 +126,7 @@ double pow(double x, double y)
 			k = (iy>>20) - 0x3ff;  /* exponent */
 			if (k > 20) {
 				j = ly>>(52-k);
-				if ((j<<(52-k)) == ly)
+				if ((j<<(52-k)) == (int32_t)ly)
 					yisint = 2 - (j&1);
 			} else if (ly == 0) {
 				j = iy>>(20-k);
